lbry: connect to ClaimTrie
For each block, all the transactions should have been validated
by the blockchain before feeding into the ClaimTrie for exchange of
the ClaimTrie Merkle Hash.

If the Merkle Hash does not match, the block is invalid.

Currently, instead of calling into the ClaimTrie API,
we save the claim script to the database.  So the ClaimTrie
can replay commands offline rapidly during the development.